+++ Laboratory/Rust/benchmarking/build.rs
@@ -8,7 +8,6 @@ const BEBOP_BIN: &str = "../../../bin/compiler/Windows-Debug/bebopc.exe";
 #[cfg(unix)]
 const BEBOP_BIN: &str = "../../../bin/compiler/Linux-Debug/bebopc";
 
-
 fn main() {
     unsafe {
         bebop::COMPILER_PATH = Some(PathBuf::from(BEBOP_BIN));
